+++ sphinx-docs/source/changelog.rst
@@ -14,7 +14,9 @@ v2.2 (TBD)
 
 - Added jupyter notebook testing to continuous integration
 - `iter_from_trajectory` can run in a user-defined time range
+- Revised to work for TF 2.4
 - Can add exclusion lists when computing nlists using htf
+
 
 *Bug Fixes*